backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
tavily-python
pypdf
SpeechRecognition
gTTS
pillow
numpy
python-dotenv
aiofiles
requests
pytest
httpx
langchain
langchain_google_genai